Add files via upload
1) the health_NoCats.csv file is a newly created, reduced file (fewer columns, only the useful ones) where the columns are now numerical not categories.

2) The MLFirstRun.ipynb file makes the conversion from categorical to numerical, creates the new .csv file, and then runs two classification routines (logistic regression, one through scikit learn, the other through StatsModels. The result of the first gives 82% accuracy at classifying the test set. The result of the second gives us information on which columns are the most useful (we can still eliminate some and it may improve the accuracy). 

3) the pdf is a printout from my Jupyter file (in case anyone has trouble running it as is). 

The nice thing about having the new non-categorical csv file is we can try different approaches to classification with that new file which is based off the old one but so much easier to work with. In the meantime, we have already the 82% accuracy which isn't bad. 

Hope you all had a nice weekend!